feat: Add initial WorkspaceSettings.xcsettings file
This commit introduces the `WorkspaceSettings.xcsettings` file to the project.
This file is currently empty but will be used to configure workspace-specific
settings for the SubTrackr project in Xcode, such as indexing preferences,
build system settings, and other developer environment configurations.
Adding this file now allows for future customization of the workspace
behavior as the project evolves.