feat: Introduce WidgetShared Swift Package for shared iOS logic
This commit adds a new 'Package.swift' file, establishing the 'WidgetShared' Swift Package Manager module.
The package is configured as a '.library' product, targeting iOS 15 and later, and utilizes Swift tools version 5.9.
This setup facilitates modularization and reusability of code, specifically for components and logic intended for integration within widgets or other parts of the application.

Signed-off-by: kallyas <kallyasmedia@gmail.com>