chore: Prioritize main app scheme in Xcode scheme list
Updated the `SubTrackr.xcodeproj/xcuserdata/tum.xcuserdatad/xcschemes/xcschememanagement.plist`
file to adjust the `orderHint` values for Xcode schemes.

Specifically, the `SubTrackr.xcscheme` `orderHint` was changed from `1` to `0`,
and the `SubTrackrWidgetExtension.xcscheme` `orderHint` was changed from `0` to `1`.

This modification ensures that the main `SubTrackr` application scheme appears as the
primary (first) option in the Xcode scheme dropdown, improving developer workflow
and default scheme selection convenience. The widget extension scheme is now
positioned as the second option.

Signed-off-by: kallyas <kallyasmedia@gmail.com>